+++ yesod-core/test/RouteLeafCompile/MissingEndpoint.hs
@@ -1,7 +1,0 @@
-{-# LANGUAGE TypeFamilies #-}
-module MissingEndpoint where
-import Yesod.Core.RouteLeaf
-import YesodCoreTest.RouteLeafHook.Foundation
-
-bad :: RouteLeaves AccountR -> ()
-bad (LeafItemR _) = ()