+++ yesod-core/test/RouteLeafCompile/NestedPattern.hs
@@ -1,7 +1,0 @@
-{-# LANGUAGE TypeFamilies #-}
-module NestedPattern where
-import Yesod.Core.RouteLeaf
-import YesodCoreTest.RouteLeafHook.Foundation
-
-bad :: RouteLeaves OrgR -> ()
-bad (AccountR _ _) = ()